Move other regression tests to test_regression
