+++ mypy.ini
@@ -1,13 +1,21 @@
 [mypy]
 plugins = pydantic.mypy
-disallow_untyped_defs = true
 # disallow_untyped_calls = true
 # disallow_untyped_decorators = false
-disallow_incomplete_defs = true
+disallow_untyped_defs = false
+disallow_untyped_calls = false
+disallow_untyped_decorators = false
+disallow_incomplete_defs = false
 check_untyped_defs = true
 exclude = (?x)(
     tests/test-local-pip/setup.py
   )
+
+[mypy-conda_lock.*]
+disallow_untyped_defs = true
+# disallow_untyped_calls = true
+# disallow_untyped_decorators = true
+disallow_incomplete_defs = true
 
 [pydantic-mypy]
 init_forbid_extra = True
@@ -36,8 +44,4 @@ ignore_missing_imports = True
 [mypy-click_default_group.*]
 ignore_missing_imports = True
 
-[mypy-test_conda_lock]
-disallow_untyped_defs = false
-disallow_untyped_calls = false
-disallow_untyped_decorators = false
-disallow_incomplete_defs = false
+[mypy-tests.*]
